Add an Aspect for user verification

+++ survey/src/main/java/com/codeit/survey/services/QuestionService.java
@@ -1,20 +1,14 @@
 package com.codeit.survey.services;
 
 import com.codeit.survey.DTOs.EntityDTOs.QuestionDTO;
+import com.codeit.survey.aspects.annotations.VerifySurveyBelongToUser;
 import com.codeit.survey.entities.Choice;
 import com.codeit.survey.entities.Question;
 import com.codeit.survey.entities.Survey;
-import com.codeit.survey.entities.SurveyUser;
 import com.codeit.survey.repositories.QuestionRepo;
-import com.codeit.survey.security.CustomUserDetails;
-import org.apache.catalina.User;
-import org.hibernate.proxy.EntityNotFoundDelegate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -26,7 +20,6 @@ public class QuestionService {
     private QuestionRepo questionRepo;
     private ChoiceService choiceService;
     private SurveyService surveyService;
-    private VerificationService verificationService;
 
     public List<QuestionDTO> createDTOsFromQuestions(List<Question> questions){
         return questions
@@ -40,11 +33,10 @@ public class QuestionService {
 
     @Autowired
     @Lazy
-    public QuestionService(QuestionRepo questionRepo, ChoiceService choiceService, SurveyService surveyService, VerificationService verificationService){
+    public QuestionService(QuestionRepo questionRepo, ChoiceService choiceService, SurveyService surveyService){
         this.questionRepo = questionRepo;
         this.choiceService = choiceService;
         this.surveyService = surveyService;
-        this.verificationService = verificationService;
     }
 
     private void setSurvey(Question question){
@@ -86,6 +78,10 @@ public class QuestionService {
 
     public ResponseEntity<?> deleteQuestionFromSurvey_admin(Integer questionId){
         try {
+            Question question = questionRepo.findById(questionId).orElse(null);
+            if (question == null){
+                return ResponseEntity.badRequest().build();
+            }
             questionRepo.deleteById(questionId);
             return ResponseEntity.ok().build();
         }catch (Exception e){
@@ -104,37 +100,17 @@ public class QuestionService {
 
 
     // user methods
+    @VerifySurveyBelongToUser
     public ResponseEntity<?> addQuestionToSurvey(Question question){
-        // do the surveys of the questions belong to the user
-        Integer surveyId = question.getSurvey().getId();
-        if ( verificationService.notUserSurvey(surveyId))
-        {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("The Question's Survey doesn't belong to the User");
-        }
-
-        // if the user has proven that he owns all the questions, he can add as an admin
         return addQuestionToSurvey_admin(question);
     }
 
     public ResponseEntity<?> deleteQuestionsFromSurvey(Integer questionId){
-        Question question = questionRepo.findById(questionId).orElse(null);
-        if (question == null){
-            return ResponseEntity.badRequest().build();
-        }
-        // do the surveys of the questions belong to the user
-        Integer surveyId = question.getSurvey().getId();
-        if (verificationService.notUserSurvey(surveyId)){
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
         return deleteQuestionFromSurvey_admin(questionId);
     }
 
+    @VerifySurveyBelongToUser
     public ResponseEntity<?> updateSurveyQuestion(Question question){
-        // do the surveys of the questions belong to the user
-        Integer surveyId = question.getSurvey().getId();
-        if (verificationService.notUserSurvey(surveyId)){
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
         return updateSurveyQuestion_admin(question);
     }
